Fix few issues caused by disabling unlimited throttling (#5427)
* Fix few issues caused by disabling unlimited throttling

This commit fixes following issues caused by disabling unlimited throttling tier in deployment.toml
	1. Invalid x-throttling tier Unlimited found error after adding and re-saving a new resource
	2. Invalid x-throttling tier Unlimited found error when deploying sample pizza-shack API

* Handle null x-throttling tier values in importOpenAPIDefinition

Null x-throttling tier values in an uploaed OpenAPI definition are replaced by default APILevelPolicy

* Handle Invalid x-throttling tier error with an error code

* Bug fixes for PR product-apim#1376

* Add default rate limiting properties to tenant-config-schema

+++ components/apimgt/org.wso2.carbon.apimgt.rest.api.admin.v1/src/main/java/org/wso2/carbon/apimgt/rest/api/admin/v1/impl/ThrottlingApiServiceImpl.java
@@ -247,6 +247,12 @@ public class ThrottlingApiServiceImpl implements ThrottlingApiService {
                     .from(ExceptionCodes.ALREADY_ASSIGNED_ADVANCED_POLICY_DELETE_ERROR,
                             existingPolicy.getPolicyName()));
         }
+        if (APIUtil.checkPolicyConfiguredAsDefault(existingPolicy.getPolicyName(),
+                PolicyConstants.POLICY_LEVEL_API, organization)) {
+            String message = "Policy " + policyId + " configured as the Default Policy.";
+            log.error(message);
+            throw new APIManagementException(message);
+        }
         apiProvider.deletePolicy(username, PolicyConstants.POLICY_LEVEL_API, existingPolicy.getPolicyName());
         return Response.ok().build();
     }
@@ -422,6 +428,12 @@ public class ThrottlingApiServiceImpl implements ThrottlingApiService {
             if (apiProvider.hasAttachments(organization, existingPolicy.getPolicyName(),
                     PolicyConstants.POLICY_LEVEL_APP,organization)) {
                 String message = "Policy " + policyId + " already attached to an application";
+                log.error(message);
+                throw new APIManagementException(message);
+            }
+            if (APIUtil.checkPolicyConfiguredAsDefault(existingPolicy.getPolicyName(),
+                    PolicyConstants.POLICY_LEVEL_APP, organization)) {
+                String message = "Policy " + policyId + " configured as the Default Policy.";
                 log.error(message);
                 throw new APIManagementException(message);
             }
@@ -692,6 +704,12 @@ public class ThrottlingApiServiceImpl implements ThrottlingApiService {
             if (apiProvider.hasAttachments(username, existingPolicy.getPolicyName(),
                     PolicyConstants.POLICY_LEVEL_SUB, organization)) {
                 String message = "Policy " + policyId + " already has subscriptions";
+                log.error(message);
+                throw new APIManagementException(message);
+            }
+            if (APIUtil.checkPolicyConfiguredAsDefault(existingPolicy.getPolicyName(),
+                    PolicyConstants.POLICY_LEVEL_SUB, organization)) {
+                String message = "Policy " + policyId + " configured as the Default Policy.";
                 log.error(message);
                 throw new APIManagementException(message);
             }
